Renders registration and login pages. Starts troubleshooting registration submission issue with Custom User model.
