--- accounts/models.py
@@ -1,9 +1,0 @@
-from django.contrib.auth.models import AbstractUser
-
-
-class CustomUser(AbstractUser):
-    pass
-    # add additional fields in here
-
-    def __str__(self):
-        return self.username
